feat: Implement undo command

+++ InitialProject/InitialProject/WPF/ViewModels/GuideViewModels/YourToursViewModel.cs
@@ -61,11 +61,14 @@ namespace InitialProject.WPF.ViewModels
 		private readonly TourReservationService _tourReservationService;
 
 		private readonly User _guide;
+
+		private readonly Stack<Tuple<ReversibleCommand, object?>> _commandStack;
         #endregion
 
-        public YourToursViewModel(User guide)
+        public YourToursViewModel(User guide, Stack<Tuple<ReversibleCommand, object?>> commandStack)
         {
 			_guide = guide;
+			_commandStack = commandStack;
 
 			_tourService = new TourService();
 			_tourReservationService = new TourReservationService();
@@ -73,7 +76,7 @@ namespace InitialProject.WPF.ViewModels
 			PastTours = new ObservableCollection<Tour>();
             FutureTours = new ObservableCollection<Tour>();
 
-			CancelTourCommand = new RelayCommand(CancelTourCommand_Execute, CancelTourCommand_CanExecute);
+			CancelTourCommand = new ReversibleCommand(CancelTourCommand_Execute, CancelTourCommand_CanExecute, CancelTourCommand_Reverse);
 
 			LoadFutureTours();
 			LoadPastTours();
@@ -97,19 +100,29 @@ namespace InitialProject.WPF.ViewModels
         }
 
         #region COMMANDS
-        public RelayCommand CancelTourCommand { get; }
+        public ReversibleCommand CancelTourCommand { get; }
         public void CancelTourCommand_Execute(object? parameter)
 		{
+			if (MessageBox.Show("Are you sure you want to cancel this tour? You can undo this action later but all the reservations will remain deleted and guests will keep their vouchers!", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Exclamation, MessageBoxResult.No) == MessageBoxResult.No) return;
 			_tourService.CancelTour(SelectedFutureTour);
 			_tourReservationService.DeleteAllReservationsForCancelledTour(SelectedFutureTour);
 			LoadFutureTours();
+
+			_commandStack.Push(new Tuple<ReversibleCommand, object?>(CancelTourCommand, parameter));
 		}
 
 		public bool CancelTourCommand_CanExecute(object? parameter)
 		{
 			Tour? tour = parameter as Tour;
-            return tour is not null && tour.Status != TourStatus.CANCELED && tour.StartTime.Subtract(DateTime.Now).TotalHours > 48;
+            return tour is not null && tour.Status != TourStatus.CANCELED && tour.StartTime.Subtract(DateTime.Now) > TimeSpan.FromDays(2);
         }
+
+		public void CancelTourCommand_Reverse(object? parameter)
+		{
+			Tour? tour = parameter as Tour;
+			_tourService.UncancelTour(tour);
+			LoadFutureTours();
+		}
         #endregion
     }
 }
